+++ src/klib/pic.hh
@@ -9,7 +9,7 @@ namespace wlib {
         void operator=(Pic const&) = delete;
 
         void static end_of_interrupt(u8 irq);
-        void static remap(u16 master_offset, u16 slave_offset);
+        void static remap(u8 master_offset, u8 slave_offset);
         void static clear_masks();
 
         auto static get() -> Pic& {
